refactor: move threjs herper to threejs sub folder
