+++ src/app/lesson_22.ts
@@ -3,7 +3,7 @@ import { GLTF, GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
 import { GUI } from "lil-gui";
 
 // HELPERS
-import initThreeJs from "../helpers/initThreeJs";
+import initThreeJs from "../helpers/threejs/initThreeJs";
 
 // MODELS
 import FoxGLTF from "../assets/models/Fox/glTF/Fox.gltf?url";
