+++ src/app/lesson_26.ts
@@ -4,7 +4,7 @@ import Cannon from "cannon";
 import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
 
 // HELPERS
-import initThreeJs from "../helpers/initThreeJs";
+import initThreeJs from "../helpers/threejs/initThreeJs";
 
 // MODELS
 import FoxGLTF from "../assets/models/Fox/glTF/Fox.gltf?url";
